Make internalparameters a pydantic model
- separate out preprocessing functions from internal param base class
- simplify internal param base classes
- implement processing in validators
- propagate json decoders from children
