+++ setup.cfg
@@ -35,7 +35,7 @@ install_requires =
     xarray
     doct
     databroker
-    dodal @ git+https://github.com/DiamondLightSource/python-dodal.git
+    dodal @ git+https://github.com/DiamondLightSource/python-dodal.git@7a92a019fb4301c3d35c782ac0dac89a55aeb0f3
 
 [options.extras_require]
 dev =
