+++ src/artemis/experiment_plans/fast_grid_scan_plan.py
@@ -42,7 +42,7 @@ if TYPE_CHECKING:
     from artemis.external_interaction.callbacks.fgs.fgs_callback_collection import (
         FGSCallbackCollection,
     )
-    from artemis.parameters.internal_parameters.plan_specific.fgs_internal_params import (
+    from artemis.parameters.plan_specific.fgs_internal_params import (
         FGSInternalParameters,
     )
 
@@ -319,7 +319,7 @@ if __name__ == "__main__":
 
     RE = RunEngine({})
     RE.waiting_hook = ProgressBarManager()
-    from artemis.parameters.internal_parameters.plan_specific.fgs_internal_params import (
+    from artemis.parameters.plan_specific.fgs_internal_params import (
         FGSInternalParameters,
     )
