+++ src/artemis/experiment_plans/full_grid_scan.py
@@ -22,13 +22,11 @@ from artemis.external_interaction.callbacks.fgs.fgs_callback_collection import (
 )
 from artemis.log import LOGGER
 from artemis.parameters.beamline_parameters import get_beamline_parameters
-from artemis.parameters.internal_parameters.plan_specific.fgs_internal_params import (
-    GridScanParams,
-)
+from artemis.parameters.plan_specific.fgs_internal_params import GridScanParams
 from artemis.utils.utils import Point3D
 
 if TYPE_CHECKING:
-    from artemis.parameters.internal_parameters.plan_specific.grid_scan_with_edge_detect_params import (
+    from artemis.parameters.plan_specific.grid_scan_with_edge_detect_params import (
         GridScanWithEdgeDetectInternalParameters,
         GridScanWithEdgeDetectParams,
     )
