+++ src/artemis/external_interaction/callbacks/fgs/ispyb_callback.py
@@ -13,9 +13,7 @@ from artemis.external_interaction.ispyb.store_in_ispyb import (
 )
 from artemis.log import LOGGER, set_dcgid_tag
 from artemis.parameters.constants import ISPYB_PLAN_NAME, SIM_ISPYB_CONFIG
-from artemis.parameters.internal_parameters.plan_specific.fgs_internal_params import (
-    FGSInternalParameters,
-)
+from artemis.parameters.plan_specific.fgs_internal_params import FGSInternalParameters
 
 
 class FGSISPyBHandlerCallback(CallbackBase):
