+++ src/artemis/external_interaction/callbacks/fgs/tests/test_ispyb_handler.py
@@ -10,9 +10,7 @@ from artemis.external_interaction.callbacks.fgs.ispyb_callback import (
 from artemis.external_interaction.callbacks.fgs.tests.conftest import TestData
 from artemis.log import LOGGER, set_up_logging_handlers
 from artemis.parameters.external_parameters import from_file as default_raw_params
-from artemis.parameters.internal_parameters.plan_specific.fgs_internal_params import (
-    FGSInternalParameters,
-)
+from artemis.parameters.plan_specific.fgs_internal_params import FGSInternalParameters
 
 DC_IDS = [1, 2]
 DCG_ID = 4
