+++ src/artemis/external_interaction/system_tests/test_ispyb_dev_connection.py
@@ -5,11 +5,8 @@ from artemis.external_interaction.ispyb.store_in_ispyb import (
     StoreInIspyb2D,
     StoreInIspyb3D,
 )
-from artemis.parameters.internal_parameters.plan_specific.fgs_internal_params import (
-    FGSInternalParameters,
-)
 from artemis.parameters.external_parameters import from_file as default_raw_params
-
+from artemis.parameters.plan_specific.fgs_internal_params import FGSInternalParameters
 
 ISPYB_CONFIG = "/dls_sw/dasc/mariadb/credentials/ispyb-dev.cfg"
