--- src/artemis/parameters/internal_parameters/__init__.py
@@ -1,5 +1,0 @@
-from artemis.parameters.internal_parameters.internal_parameters import (
-    InternalParameters,
-)
-
-__all__ = ["InternalParameters"]
