+++ src/artemis/parameters/schemas/full_external_parameters_schema.json
@@ -3,7 +3,7 @@
     "type": "object",
     "properties": {
         "params_version": {
-            "const": 0.4
+            "const": "0.0.5"
         },
         "artemis_params": {
             "type": "object",
