+++ src/artemis/parameters/tests/test_data/bad_test_parameters_wrong_version.json
@@ -1,5 +1,5 @@
 {
-    "params_version": 0.0,
+    "params_version": "0.0.4",
     "artemis_params": {
         "beamline": "BL03S",
         "detector": "EIGER2_X_16M",
