+++ src/artemis/parameters/tests/test_data/good_test_parameters.json
@@ -1,5 +1,5 @@
 {
-    "params_version": 0.4,
+    "params_version": "0.0.5",
     "artemis_params": {
         "beamline": "BL03S",
         "detector": "EIGER2_X_16M",
