+++ test_parameter_defaults.json
@@ -1,5 +1,5 @@
 {
-    "params_version": 0.4,
+    "params_version": "0.0.5",
     "artemis_params": {
         "zocalo_environment": "dev_artemis",
         "beamline": "BL03S",
